MooseFS 3.0.74-1
 * (master) fixed rebalance replication (check for all chunk copies for destination - not only valid)
 * (master+mount) new mechanism for atime+mtime setting during I/O
 * (master+mount) new I/O synchronization between different mounts (with cache invalidation)
 * (master+mount) new chunk number/version cache (with automatic invalidation from master)
 * (master) added mapping chunkserver IP classes (allow to have separate network for I/O and separate for other activity)
 * (master) fixed status returned by writechunk after network down/up
 * (master) changed trashtime from seconds to hours
 * (master) added METADATA_SAVE_FREQ option (allow to save metadata less frequently than every hour)
 * (master) added using in emergency (endangered chunks) all available servers for replication (even overloaded and being maintained)
 * (master) added using chunkservers in 'internal rebalance' state in case of deleting chunks
 * (all) spell check errors fixed (patch committed by Dmitry Smirnov)

+++ mfscommon/massert.h
@@ -33,30 +33,30 @@
 
 #include "strerr.h"
 
-#define massert(e,msg) ((e) ? (void)0 : (fprintf(stderr,"%s:%u - failed assertion '%s' : %s\n",__FILE__,__LINE__,#e,(msg)),syslog(LOG_ERR,"%s:%u - failed assertion '%s' : %s",__FILE__,__LINE__,#e,(msg)),abort()))
-#define sassert(e) ((e) ? (void)0 : (fprintf(stderr,"%s:%u - failed assertion '%s'\n",__FILE__,__LINE__,#e),syslog(LOG_ERR,"%s:%u - failed assertion '%s'",__FILE__,__LINE__,#e),abort()))
+#define massert(e,msg) ((e) ? (void)0 : (fprintf(stderr,"%s:%u - failed assertion '%s' : %s\n",__FILE__,(unsigned)__LINE__,#e,(msg)),syslog(LOG_ERR,"%s:%u - failed assertion '%s' : %s",__FILE__,(unsigned)__LINE__,#e,(msg)),abort()))
+#define sassert(e) ((e) ? (void)0 : (fprintf(stderr,"%s:%u - failed assertion '%s'\n",__FILE__,(unsigned)__LINE__,#e),syslog(LOG_ERR,"%s:%u - failed assertion '%s'",__FILE__,(unsigned)__LINE__,#e),abort()))
 #define passert(ptr) if (ptr==NULL) { \
-		fprintf(stderr,"%s:%u - out of memory: %s is NULL\n",__FILE__,__LINE__,#ptr); \
-		syslog(LOG_ERR,"%s:%u - out of memory: %s is NULL",__FILE__,__LINE__,#ptr); \
+		fprintf(stderr,"%s:%u - out of memory: %s is NULL\n",__FILE__,(unsigned)__LINE__,#ptr); \
+		syslog(LOG_ERR,"%s:%u - out of memory: %s is NULL",__FILE__,(unsigned)__LINE__,#ptr); \
 		abort(); \
 	} else if (ptr==((void*)(-1))) { \
 		const char *_mfs_errorstring = strerr(errno); \
-		syslog(LOG_ERR,"%s:%u - mmap error on %s, error: %s",__FILE__,__LINE__,#ptr,_mfs_errorstring); \
-		fprintf(stderr,"%s:%u - mmap error on %s, error: %s\n",__FILE__,__LINE__,#ptr,_mfs_errorstring); \
+		syslog(LOG_ERR,"%s:%u - mmap error on %s, error: %s",__FILE__,(unsigned)__LINE__,#ptr,_mfs_errorstring); \
+		fprintf(stderr,"%s:%u - mmap error on %s, error: %s\n",__FILE__,(unsigned)__LINE__,#ptr,_mfs_errorstring); \
 		abort(); \
 	}
 #define eassert(e) if (!(e)) { \
 		const char *_mfs_errorstring = strerr(errno); \
-		syslog(LOG_ERR,"%s:%u - failed assertion '%s', error: %s",__FILE__,__LINE__,#e,_mfs_errorstring); \
-		fprintf(stderr,"%s:%u - failed assertion '%s', error: %s\n",__FILE__,__LINE__,#e,_mfs_errorstring); \
+		syslog(LOG_ERR,"%s:%u - failed assertion '%s', error: %s",__FILE__,(unsigned)__LINE__,#e,_mfs_errorstring); \
+		fprintf(stderr,"%s:%u - failed assertion '%s', error: %s\n",__FILE__,(unsigned)__LINE__,#e,_mfs_errorstring); \
 		abort(); \
 	}
 #define zassert(e) { \
 	int _mfs_assert_ret = (e); \
 	if (_mfs_assert_ret!=0) { \
 		const char *_mfs_errorstring = strerr(errno); \
-		syslog(LOG_ERR,"%s:%u - unexpected status, '%s' returned: %d (errno: %s)",__FILE__,__LINE__,#e,_mfs_assert_ret,_mfs_errorstring); \
-		fprintf(stderr,"%s:%u - unexpected status, '%s' returned: %d (errno: %s)\n",__FILE__,__LINE__,#e,_mfs_assert_ret,_mfs_errorstring); \
+		syslog(LOG_ERR,"%s:%u - unexpected status, '%s' returned: %d (errno: %s)",__FILE__,(unsigned)__LINE__,#e,_mfs_assert_ret,_mfs_errorstring); \
+		fprintf(stderr,"%s:%u - unexpected status, '%s' returned: %d (errno: %s)\n",__FILE__,(unsigned)__LINE__,#e,_mfs_assert_ret,_mfs_errorstring); \
 		abort(); \
 	} \
 }
